apps/nl2sql/requirements.txt
fastapi
uvicorn
pyodbc
google-generativeai
python-dotenv
pandas
openpyxl
jinja2
python-multipart
aiofiles